+++ gateway/src/main/java/org/acumos/federation/gateway/service/impl/CatalogServiceImpl.java
@@ -37,11 +37,13 @@ import javax.annotation.PostConstruct;
 import org.acumos.cds.AccessTypeCode;
 import org.acumos.cds.ValidationStatusCode;
 import org.acumos.cds.client.ICommonDataServiceRestClient;
+import org.acumos.cds.domain.MLPDocument;
 import org.acumos.cds.domain.MLPArtifact;
 import org.acumos.cds.domain.MLPSolution;
 import org.acumos.cds.domain.MLPSolutionRevision;
 import org.acumos.cds.transport.RestPageRequest;
 import org.acumos.cds.transport.RestPageResponse;
+import org.acumos.federation.gateway.cds.Document;
 import org.acumos.federation.gateway.cds.Artifact;
 import org.acumos.federation.gateway.cds.Solution;
 import org.acumos.federation.gateway.cds.SolutionRevision;
@@ -148,20 +150,6 @@ public class CatalogServiceImpl extends AbstractServiceImpl
 																					null,	//authorKeywords
 																					null, //publisherKeywords
 																					pageRequest);
-/*
-	RestPageResponse<MLPSolution> findPortalSolutions(
-																					String[] nameKeywords,
-																					String[] descriptionKeywords,
-																					boolean active,
-																					String[] userIds,
-																					String[] accessTypeCodes,
-																					String[] modelTypeCodes,
-																					String[] validationStatusCodes,
-																					String[] tags,
-																					String[] authorKeywords,
-																					String[] publisherKeywords,
-			RestPageRequest pageRequest);
-*/
 					pageSolutions = pageResponse.getContent();
 				}
 				log.debug(EELFLoggerDelegate.debugLogger, "getSolutions page response {}", pageResponse);
@@ -246,7 +234,16 @@ public class CatalogServiceImpl extends AbstractServiceImpl
 		try {
 			SolutionRevision revision =
 					(SolutionRevision)cdsClient.getSolutionRevision(theSolutionId, theRevisionId);
-			revision.setArtifacts(cdsClient.getSolutionRevisionArtifacts(theSolutionId, theRevisionId));
+			revision.setArtifacts(getSolutionRevisionArtifacts(theSolutionId, theRevisionId, theContext.withAttribute(Attributes.cdsClient, cdsClient)));
+			revision.setDocuments(getSolutionRevisionDocuments(theSolutionId, theRevisionId, theContext.withAttribute(Attributes.cdsClient, cdsClient)));
+			try {
+				revision.setRevisionDescription(cdsClient.getRevisionDescription(theRevisionId, AccessTypeCode.PB.name()));
+			}
+			catch (HttpStatusCodeException restx) {
+				if (!Errors.isCDSNotFound(restx))
+					throw new ServiceException("Failed to retrieve solution revision description", restx);
+			}
+	
 			return revision;
 		}	
 		catch (HttpStatusCodeException restx) {
@@ -263,13 +260,13 @@ public class CatalogServiceImpl extends AbstractServiceImpl
 
 		log.trace(EELFLoggerDelegate.debugLogger, "getSolutionRevisionArtifacts");
 		try {
-			return getClient().getSolutionRevisionArtifacts(theSolutionId, theRevisionId);
+			return getClient(theContext).getSolutionRevisionArtifacts(theSolutionId, theRevisionId);
 		}
 		catch (HttpStatusCodeException restx) {
 			if (Errors.isCDSNotFound(restx))
 				return null;
 			else
-				throw new ServiceException("Failed to retrieve solution information", restx);
+				throw new ServiceException("Failed to retrieve solution revision artifacts information", restx);
 		}
 	}
 
@@ -282,6 +279,7 @@ public class CatalogServiceImpl extends AbstractServiceImpl
 
 		log.trace(EELFLoggerDelegate.debugLogger, "getSolutionRevisionArtifact");
 		try {
+			//one should check that this belongs to at least one public revision of some solution accessible within the given context ..
 			return (Artifact)getClient().getArtifact(theArtifactId);
 		}	
 		catch (HttpStatusCodeException restx) {
@@ -289,6 +287,35 @@ public class CatalogServiceImpl extends AbstractServiceImpl
 				return null;
 			else
 				throw new ServiceException("Failed to retrieve solution revision artifact information", restx);
+		}
+	}
+
+	@Override
+	public List<MLPDocument> getSolutionRevisionDocuments(String theSolutionId, String theRevisionId, ServiceContext theContext) throws ServiceException {
+		log.trace(EELFLoggerDelegate.debugLogger, "getSolutionRevisionDocuments");
+		try {
+			return getClient(theContext).getSolutionRevisionDocuments(theRevisionId, AccessTypeCode.PB.name());
+		}
+		catch (HttpStatusCodeException restx) {
+			if (Errors.isCDSNotFound(restx))
+				return null;
+			else
+				throw new ServiceException("Failed to retrieve solution revision documents information", restx);
+		}
+	}
+
+	@Override
+	public Document getSolutionRevisionDocument(String theDocumentId, ServiceContext theContext) throws ServiceException {
+		log.trace(EELFLoggerDelegate.debugLogger, "getSolutionRevisionDocument");
+		try {
+			//one should check that this has a public visibility within at least one revision of some solution accessible within the given context ..
+			return (Document)getClient().getDocument(theDocumentId);
+		}	
+		catch (HttpStatusCodeException restx) {
+			if (Errors.isCDSNotFound(restx))
+				return null;
+			else
+				throw new ServiceException("Failed to retrieve solution revision document information", restx);
 		}
 	}